Extract date_range property from change sets into a mixin
refs #2473
