+++ app/resources/numismatic_issues/numismatic_issue_change_set.rb
@@ -4,9 +4,9 @@ class NumismaticIssueChangeSet < ChangeSet
   apply_workflow(DraftCompleteWorkflow)
 
   include VisibilityProperty
+  include DateRangeProperty
   property :artist, multiple: true, required: false, default: []
   property :color, multiple: false, required: false
-  property :date_range, multiple: false, required: false
   property :denomination, multiple: false, required: false
   property :description, multiple: false, required: false
   property :edge, multiple: false, required: false
@@ -59,7 +59,6 @@ class NumismaticIssueChangeSet < ChangeSet
 
   # Virtual Attributes
   property :files, virtual: true, multiple: true, required: false
-  property :date_range_form_attributes, virtual: true
 
   validates_with AutoIncrementValidator, property: :issue_number
   validates_with CollectionValidator
@@ -69,26 +68,6 @@ class NumismaticIssueChangeSet < ChangeSet
   validates_with ViewingDirectionValidator
   validates_with ViewingHintValidator
   validates :visibility, presence: true
-  validate :date_range_validity
-
-  def date_range_form_attributes=(attributes)
-    return unless date_range_form.validate(attributes)
-    date_range_form.sync
-    self.date_range = date_range_form.model
-  end
-
-  def date_range_validity
-    return if date_range_form.valid?
-    errors.add(:date_range_form, "is not valid.")
-  end
-
-  def date_range_form
-    @date_range_form ||= DynamicChangeSet.new(date_range_value || DateRange.new).tap(&:prepopulate!)
-  end
-
-  def date_range_value
-    Array.wrap(date_range).first
-  end
 
   def primary_terms
     {
